outbox & cap

+++ src/BuildingBlocks/BuildingBlocks.Abstractions/Domain/IHaveCreator.cs
@@ -3,5 +3,5 @@ namespace BuildingBlocks.Abstractions.Domain;
 public interface IHaveCreator
 {
     DateTime Created { get; }
-    int? CreatedBy { get; }
+    long? CreatedBy { get; }
 }
